Split tutorial code
